% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_cache.R
\name{ai_cache}
\alias{ai_cache}
\title{Download an updated list of country, indicator, and source information}
\usage{
ai_cache(lang = c("en", "es", "fr", "ar", "zh"))
}
\arguments{
\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default.}
}
\value{
A list containing the following items:
}
\description{
Download an updated list of information regarding countries, indicators
from the World Bank API, Numbers for Development API and No Celings
}
\note{
Not all data returns have support for langauges other than english. If the specific return
does not support your requested language by default it will return \code{NA}.
The options for \code{lang} are:
\itemize{
\item \code{en}: English
\item \code{es}: Spanish
\item \code{fr}: French
\item \code{ar}: Arabic
\item \code{zh}: Mandarin
}

Saving this return and using it has the \code{cache} parameter in \code{\link{ai_search}} and \code{\link{ind_search}}
replaces the default cached version \code{\link{ai_cachelist}} that comes with the package itself
}
\examples{
# default is english. To specific another language use argument lang
ai_cache(lang = "es")
}
