% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_GovData360.R
\name{load.360.metadata}
\alias{load.360.metadata}
\title{Download updated indicator metadata from Govdata360 platform govdata360.worldbank.org
Only is english.}
\usage{
load.360.metadata()
}
\value{
A data frame with the metadata of all the indicators of Govdata360
}
\description{
Download updated indicator metadata from Govdata360 platform govdata360.worldbank.org
Only is english.
}
\examples{
# 
load.360.medatada()
}
