% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_N4D.R
\name{load.N4D.data}
\alias{load.N4D.data}
\title{Download Data from the Numbers for Development (N4D)}
\usage{
load.N4D.data(pIndicators, pCountry = "all", pStart = 2010, pEnd = 2015,
  cache)
}
\arguments{
\item{cache}{List of data frames returned from \code{\link{ai_cache}}. If omitted,
\code{\link{ai)_cachelist}} is used}

\item{country}{Character vector of country or region codes. Default value is special code of \code{all}.
Other permissible values are codes in the following fields from the \code{\link{ai_cachelist}} \code{countries_idb}
data frame.  \code{iso2c}}

\item{indicator}{Character vector of indicator codes. These codes correspond to the \code{src_id_ind} column
from the \code{indicator} data frame of \code{\link{ai_cache}} or \code{\link{ai_cachelist}}, or
the result of \code{\link{ai_search}}}

\item{startdate}{Numeric. Year (four digit) of the start of the requested date range.}

\item{enddate}{Numeric. Year (four digit) of the end of the requested date range.}

\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default.}
}
\value{
Data frame with all available requested data.
}
\description{
This function downloads the requested information using the Numbers for Development API
}
\note{
Not all data returns have support for langauges other than english. If the specific return
 does not support your requested language by default it will return \code{NA}.
 The options for \code{lang} are:
 \itemize{
 \item \code{en}: English
 \item \code{es}: Spanish
 \item \code{fr}: French
 \item \code{ar}: Arabic
 \item \code{zh}: Mandarin
 }
 If there is no data available that matches the request parameters, an empty data frame is returned along with a
 \code{warning}.

 @examples
load.NC.data(pIndicators=c("SOC_046"))
load.N4D.data(pIndicators=c("SOC_046"),pStart=2013,pEnd=2015, pCountry='all')
}
