% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_cache.R
\name{meta_indicators}
\alias{meta_indicators}
\title{Download updated indicator information from World Bank API, Numbers for Development API and No Celings}
\usage{
meta_indicators(lang = c("en", "es"))
}
\arguments{
\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default.}
}
\value{
A data frame of available indicators with related information
}
\description{
Download updated information on available indicators
from from World Bank API, Numbers for Development API and No Celings
}
\note{
Not all data returns have support for langauges other than english. If the specific return
does not support your requested language by default it will return \code{NA}. For an enumeration of
supported languages by data source please see TODO.
The options for \code{lang} are:
\itemize{
\item \code{en}: English
\item \code{es}: Spanish
}
}
\examples{
# default is english. To specific another language use argument lang
meta_indicators(lang = "es")
}
