% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_NCD.R
\name{load.NC.metadata}
\alias{load.NC.metadata}
\title{Download updated indicator metadata from Numbers for Development (N4D) API}
\usage{
load.NC.metadata(lang = c("en", "es"))
}
\arguments{
\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default. This source is only available in english. The only column supports other languages is the topic}
}
\value{
A data frame of available indicators with related metadata
}
\description{
Download updated indicator metadata from Numbers for Development (N4D) API
}
\note{
The only language supported for this source of information is english
}
\examples{
load.NC.metadata()
}
